Created new target 'neomouseTypes', TODO move all types to there to be shared across all other targets

+++ Sources/neomouseConfig/config.swift
@@ -28,7 +28,7 @@ public struct Config: Decodable, Sendable {
     }
 
     public struct Motion: Decodable, Sendable {
-        public let linesOnScreen: CGFloat
+        public let linesOnScreen: Int
         public let rangeX: CGFloat
         public let rangeY: CGFloat
         public let isClampCursorToCurrentScreen: Bool
